Move transform into general axes widget
