+++ src/napari_metadata/_axes_widget.py
@@ -1,6 +1,14 @@
-from typing import TYPE_CHECKING, List, Tuple
+from typing import TYPE_CHECKING, List, Optional, Tuple
 
-from qtpy.QtWidgets import QComboBox, QGridLayout, QLabel, QLineEdit, QWidget
+from qtpy.QtWidgets import (
+    QAbstractSpinBox,
+    QComboBox,
+    QDoubleSpinBox,
+    QGridLayout,
+    QLabel,
+    QLineEdit,
+    QWidget,
+)
 
 from napari_metadata._axis_type import AxisType
 from napari_metadata._model import (
@@ -23,14 +31,26 @@ if TYPE_CHECKING:
     from napari.layers import Layer
 
 
-class AxisNameTypeRow:
+def make_double_spinbox(value: float, *, lower: float) -> QDoubleSpinBox:
+    spinbox = QDoubleSpinBox()
+    spinbox.setDecimals(6)
+    spinbox.setMinimum(lower)
+    spinbox.setValue(value)
+    spinbox.setSingleStep(0.1)
+    spinbox.setButtonSymbols(QAbstractSpinBox.ButtonSymbols.NoButtons)
+    return spinbox
+
+
+class AxisRow:
     def __init__(self) -> None:
         self.name: QLineEdit = QLineEdit()
         self.type: QComboBox = QComboBox()
         self.type.addItems(AxisType.names())
+        self.scale = make_double_spinbox(1, lower=1e-6)
+        self.translate = make_double_spinbox(0, lower=-1e6)
 
     def widgets(self) -> Tuple[QWidget, ...]:
-        return (self.name, self.type)
+        return (self.name, self.type, self.scale, self.translate)
 
     def to_axis(
         self,
@@ -46,19 +66,22 @@ class AxisNameTypeRow:
         return SpaceAxis(name=self.name.text(), unit=space_unit)
 
 
-class AxesNameTypeWidget(QWidget):
+class AxesWidget(QWidget):
     """Shows and controls all axes' names and types."""
 
     def __init__(self, viewer: "ViewerModel") -> None:
         super().__init__()
         self._viewer: "ViewerModel" = viewer
-        self._rows: List[AxisNameTypeRow] = []
+        self._layer: Optional["Layer"] = None
+        self._rows: List[AxisRow] = []
 
         layout = QGridLayout()
         layout.setContentsMargins(0, 0, 0, 0)
 
         layout.addWidget(QLabel("Name"), 0, 0)
         layout.addWidget(QLabel("Type"), 0, 1)
+        layout.addWidget(QLabel("Scale"), 0, 2)
+        layout.addWidget(QLabel("Translate"), 0, 3)
 
         self.setLayout(layout)
 
@@ -85,6 +108,20 @@ class AxesNameTypeWidget(QWidget):
             dims.axis_labels = names
         ndim_diff = dims.ndim - layer.ndim
 
+        old_layer = self._layer
+        if old_layer is not None:
+            old_layer.events.scale.disconnect(self._on_layer_scale_changed)
+            old_layer.events.translate.disconnect(
+                self._on_layer_translate_changed
+            )
+        if layer is not None:
+            layer.events.scale.connect(self._on_layer_scale_changed)
+            layer.events.translate.connect(self._on_layer_translate_changed)
+        self._layer = layer
+        if self._layer is not None:
+            self._on_layer_scale_changed()
+            self._on_layer_translate_changed()
+
         extras = coerce_extra_metadata(self._viewer, layer)
         for i, row in enumerate(self._rows):
             set_row_visible(row, i >= ndim_diff)
@@ -99,10 +136,12 @@ class AxesNameTypeWidget(QWidget):
         viewer_names[-layer.ndim :] = layer_names  # noqa
         return tuple(viewer_names)
 
-    def _make_row(self) -> AxisNameTypeRow:
-        widget = AxisNameTypeRow()
+    def _make_row(self) -> AxisRow:
+        widget = AxisRow()
         widget.name.textChanged.connect(self._on_axis_name_changed)
         widget.type.currentTextChanged.connect(self._on_axis_type_changed)
+        widget.scale.valueChanged.connect(self._on_pixel_size_changed)
+        widget.translate.valueChanged.connect(self._on_translate_changed)
         return widget
 
     def _on_viewer_dims_axis_labels_changed(self) -> None:
@@ -115,8 +154,37 @@ class AxesNameTypeWidget(QWidget):
             axis_names = self._viewer.dims.axis_labels[-layer.ndim :]  # noqa
             extras.set_axis_names(axis_names)
 
-    def axis_widgets(self) -> Tuple[AxisNameTypeRow, ...]:
+    def _on_layer_scale_changed(self) -> None:
+        assert self._layer is not None
+        scale = self._layer.scale
+        widgets = self._layer_widgets()
+        for s, w in zip(scale, widgets):
+            w.scale.setValue(s)
+
+    def _on_layer_translate_changed(self) -> None:
+        assert self._layer is not None
+        translate = self._layer.translate
+        widgets = self._layer_widgets()
+        for t, w in zip(translate, widgets):
+            w.translate.setValue(t)
+
+    def _on_pixel_size_changed(self) -> None:
+        scale = tuple(w.scale.value() for w in self._layer_widgets())
+        self._layer.scale = scale
+
+    def _on_translate_changed(self) -> None:
+        translate = tuple(w.translate.value() for w in self._layer_widgets())
+        self._layer.translate = translate
+
+    def axis_widgets(self) -> Tuple[AxisRow, ...]:
         return tuple(self._rows)
+
+    def _layer_widgets(self) -> Tuple[AxisRow, ...]:
+        return (
+            ()
+            if self._layer is None
+            else tuple(self._rows[-self._layer.ndim :])  # noqa
+        )
 
     def axis_names(self) -> Tuple[str, ...]:
         return tuple(widget.name.text() for widget in self.axis_widgets())
@@ -139,27 +207,32 @@ class AxesNameTypeWidget(QWidget):
                 )
 
 
-class ReadOnlyAxisNameTypeRow:
+class ReadOnlyAxisRow:
     def __init__(self) -> None:
         self.name: QLineEdit = readonly_lineedit()
         self.type: QLineEdit = readonly_lineedit()
+        self.scale: QLineEdit = readonly_lineedit()
+        self.translate: QLineEdit = readonly_lineedit()
 
     def widgets(self) -> Tuple[QWidget, ...]:
-        return (self.name, self.type)
+        return (self.name, self.type, self.scale, self.translate)
 
 
-class ReadOnlyAxesNameTypeWidget(QWidget):
+class ReadOnlyAxesWidget(QWidget):
     """Shows all axes' names and types."""
 
     def __init__(self, viewer: "ViewerModel") -> None:
         super().__init__()
         self._viewer: "ViewerModel" = viewer
-        self._rows: List[ReadOnlyAxisNameTypeRow] = []
+        self._layer: Optional["Layer"] = None
+        self._rows: List[ReadOnlyAxisRow] = []
 
         layout = QGridLayout()
         layout.setContentsMargins(0, 0, 0, 0)
         layout.addWidget(QLabel("Name"), 0, 0)
         layout.addWidget(QLabel("Type"), 0, 1)
+        layout.addWidget(QLabel("Scale"), 0, 2)
+        layout.addWidget(QLabel("Translate"), 0, 3)
         self.setLayout(layout)
 
         self._viewer.dims.events.axis_labels.connect(
@@ -172,7 +245,7 @@ class ReadOnlyAxesNameTypeWidget(QWidget):
             rows=self._rows,
             layout=self.layout(),
             desired_num=dims.ndim,
-            row_factory=ReadOnlyAxisNameTypeRow,
+            row_factory=ReadOnlyAxisRow,
         )
         ndim_diff = dims.ndim - layer.ndim
         extras = coerce_extra_metadata(self._viewer, layer)
@@ -182,6 +255,39 @@ class ReadOnlyAxesNameTypeWidget(QWidget):
                 axis = extras.axes[i - ndim_diff]
                 row.name.setText(axis.name)
                 row.type.setText(str(axis.get_type()))
+
+        old_layer = self._layer
+        if old_layer is not None:
+            old_layer.events.scale.disconnect(self._on_layer_scale_changed)
+            old_layer.events.translate.disconnect(
+                self._on_layer_translate_changed
+            )
+        if layer is not None:
+            layer.events.scale.connect(self._on_layer_scale_changed)
+            layer.events.translate.connect(self._on_layer_translate_changed)
+        self._layer = layer
+        if self._layer is not None:
+            self._on_layer_scale_changed()
+            self._on_layer_translate_changed()
+
+    def _on_layer_scale_changed(self) -> None:
+        assert self._layer is not None
+        scale = self._layer.scale
+        for s, r in zip(scale, self._layer_rows()):
+            r.scale.setText(str(s))
+
+    def _on_layer_translate_changed(self) -> None:
+        assert self._layer is not None
+        translate = self._layer.translate
+        for t, r in zip(translate, self._layer_rows()):
+            r.translate.setText(str(t))
+
+    def _layer_rows(self) -> Tuple[ReadOnlyAxisRow, ...]:
+        return (
+            ()
+            if self._layer is None
+            else tuple(self._rows[-self._layer.ndim :])  # noqa
+        )
 
     def _on_viewer_dims_axis_labels_changed(self) -> None:
         self._set_axis_names(self._viewer.dims.axis_labels)
